switch to u64 for publication store

+++ mqtt/mqtt-bridge/src/persist/loader.rs
@@ -185,7 +185,7 @@ mod tests {
         let num_elements = 10 as usize;
         for i in 0..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let key = Key { offset: i as u32 };
+            let key = Key { offset: i as u64 };
             let publication = Publication {
                 topic_name: i.to_string(),
                 qos: QoS::ExactlyOnce,
@@ -203,7 +203,7 @@ mod tests {
 
         for count in 0..num_elements {
             #[allow(clippy::cast_possible_truncation)]
-            let num_elements = count as u32;
+            let num_elements = count as u64;
 
             assert_eq!(elements.pop_front().unwrap().0.offset, num_elements)
         }
